refactoring app
