+++ src/components/app/index.js
@@ -1,0 +1,3 @@
+import App from './app';
+
+export default App;
